+++ test/widget/life_road_test.dart
@@ -14,7 +14,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:provider/provider.dart';
 
 import '../mocks/dice.dart';
-import 'helper/widget_build_helper.dart';
+import 'helper/testable_app.dart';
 
 Future<void> main() async {
   final i18n = await I18n.load(const Locale('en', 'US'));
