+++ test/widget/player_action_test.dart
@@ -16,7 +16,7 @@ import 'package:provider/provider.dart';
 
 import '../mocks/dice.dart';
 import '../mocks/play_room_state.dart';
-import 'helper/widget_build_helper.dart';
+import 'helper/testable_app.dart';
 
 Future<void> main() async {
   final i18n = await I18n.load(const Locale('en', 'US'));
